FROMLIST: scsi: ufs: atomic update for clkgating_enable
When giving a stress test which enables/disables clkgating, we hit device
timeout sometimes. This patch avoids subtle racy condition to address it.

Note that, this requires a patch to address the device stuck by REQ_CLKS_OFF in
__ufshcd_release().

The fix is "scsi: ufs: avoid to call REQ_CLKS_OFF to CLKS_OFF".

Link: https://lore.kernel.org/linux-scsi/20201028194352.GA3060274@google.com/T/#m979bdb55500f3f63fa4e975524f8559c80697862
Signed-off-by: Jaegeuk Kim <jaegeuk@google.com>
Change-Id: Ice47cbd199850aadd15166ac414fd97db3a79aad
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>