Revert "cpufreq: schedutil: Ignore rate limit when scaling up with FIE present"
Commit "cpufreq: schedutil: Fix superfluous updates caused by
need_freq_update" will receive a fix with the upcoming commit
"cpufreq/sched: Fix the usage of CPUFREQ_NEED_UPDATE_LIMITS".

The upcoming commit needs to be adapted for this current commit
which we currently lack. So, revert it and reimport with proper
fixes afterwards.

This reverts commit f64984260f69b6f1538c634823e1f91be2dcd2a5.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>